Refactored csv validation on backend into shared folder
