+++ packages/api/prod.Dockerfile
@@ -2,11 +2,13 @@ FROM node:16.15.0
 
 WORKDIR /usr/src/app
 
-COPY ./package*.json ./
+COPY ./api/package*.json ./usr/src/app/
 
 RUN npm install
 
-COPY . .
+COPY ./api/ /usr/src/app/
+
+COPY ./shared/ /usr/src/shared/
 
 RUN npm install -g nodemon
