+++ packages/shared/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "shared",
+  "type": "module"
+}
