+++ packages/webapp/prod.Dockerfile
@@ -2,11 +2,11 @@ FROM node:16.13.2 as build
 
 WORKDIR /usr/src/app
 
-COPY ./package.json ./.npmrc ./pnpm-lock.yaml /usr/src/app/
+COPY ./webapp/package.json ./webapp/.npmrc ./webapp/pnpm-lock.yaml /usr/src/app/
 
 RUN npm install pnpm -g && pnpm install
 
-COPY ./ /usr/src/app/
+COPY ./webapp/ /usr/src/app/
 
 RUN pnpm run build
